+++ lib/src/core/Main.cpp
@@ -6,6 +6,8 @@
 #include "./Backup.h"
 #include "Version.h"
 #include "../core/ZipCompressor.h"
+namespace {
+}
 int main(const int argc, const char* argv[]) {
   std::cout << "Sequoia v" << SEQUOIA_VERSION_MAJOR << "." << SEQUOIA_VERSION_MINOR << "." << SEQUOIA_VERSION_PATCH << " " << SEQUOIA_VERSION_CHANNEL << std::endl;
   if (argc <= 1) {
@@ -53,7 +55,7 @@ int main(const int argc, const char* argv[]) {
   std::string worldPath = "";
   
   
-  for (int i = 2; i < argc-1; i++) {
+  for (int i = 2; i < argc; i++) {
     if ((std::string(argv[i]) == "-w" || std::string(argv[i]) == "--world") && worldPath == "") {
       // then our world is here
       worldPath = std::string(argv[i+1]);
@@ -76,14 +78,78 @@ int main(const int argc, const char* argv[]) {
   std::cout << "INFO: opened world " << worldPath << std::endl;
   
   if (option == "backups") {
+    std::vector<sequoia::Backup> backups = world.getBackups();
     std::string command = argv[2];
-    
-    if (command == "view") {
-      std::vector<sequoia::Backup> backups = world.getBackups();
+    if (command == "restore") {
+
       if (backups.size() == 0) {
+        std::cout << "no backups to restore! create a backup, then return to here to restore." << std::endl;
+      } else {
+        std::cout << "select backup to restore: " << std::endl;
+        int i = 1;
+        
+        for (sequoia::Backup bc : backups) {
+          std::chrono::system_clock::time_point tp =
+            std::chrono::system_clock::from_time_t(bc.getBackupTime());
+          auto tpS = std::chrono::floor<std::chrono::seconds>(tp);
+
+          std::string formatted = std::format(
+            "{:%b %d %Y %I:%M:%S %p}", tpS);
+          
+          
+          std::cout << "  [" << i++ << "] " << formatted << " (" << bc.getRelativeLocation().substr(1) << ")" << std::endl;
+        }
+        std::string option;
+        std::cin >> option;
+
+        std::cout << "create backup before restoring? it is HIGHLY advised to backup your data before restoring to avoid data loss. (y/n)";
+
+        std::string backupConfirmation;
+        std::cin >> backupConfirmation;
+        std::transform(backupConfirmation.begin(), backupConfirmation.end(), backupConfirmation.begin(), [](unsigned char c) {
+          return std::tolower(c);
+        });
+        if (backupConfirmation == "y") {
+          auto conf = world.getBackupConfig();
+          if (!conf.has_value()) {
+            std::cerr << "error: no backup config! rerun with backups create for backup config templates!" << std::endl;
+            return 0;
+          }
+          sequoia::Backup bc(world, conf.value());
+          std::cout << "starting backup!" << std::endl;
+          bc.backup();
+          std::cout << "backed up world!" << std::endl;
+        }
+        sequoia::Backup bc = backups[std::stoi(option) - 1];
+        bc.restore();
+        std::cout << "restored world!" << std::endl;
+        return 0;
+      }
+      
+    }
+    if (command == "view") {
+      
+      if (backups.size() == 0) {
+
         std::cout << "backups:" << std::endl;
         std::cout << "  no backups, backup one and it will appear here!" << std::endl;
+        return 0;
+      }
+      else {
+        std::cout << "backups:" << std::endl;
+        for (sequoia::Backup bc : backups) {
+          std::chrono::system_clock::time_point tp =
+            std::chrono::system_clock::from_time_t(bc.getBackupTime());
+          auto tpS = std::chrono::floor<std::chrono::seconds>(tp);
 
+          std::string formatted = std::format(
+            "{:%b %d %Y %I:%M:%S %p}", tpS);
+
+
+          std::cout << "  "<< formatted << " (" << bc.getRelativeLocation().substr(1) << ")" << std::endl;
+          
+        }
+        return 0;
       }
     }
     if (command == "create") {
@@ -123,11 +189,11 @@ int main(const int argc, const char* argv[]) {
       return 0;
     }
     else {
-      std::cout << "ERROR: please select command. possible options are: view, create, delete, config" << std::endl;
+      std::cout << "ERROR: please select command. possible options are: view, create, delete, config, restore" << std::endl;
     }
   }
   else {
-    std::cout << "ERROR: please select option. possible options are: backups[view|create|delete|config], init, deinit" << std::endl;
+    std::cout << "ERROR: please select option. possible options are: backups[view|create|delete|config|restore], init, deinit" << std::endl;
   }
   
   return 0;